review: update readme and move tests
