+++ bin/stress-test/src/seeding/metrics.rs
