+++ bin/stress-test/src/store/metrics.rs
